why is text export in .dmf
